Add users controller actions

+++ episodic-shows/src/main/resources/db/migration/V3__CREATE_EPISODES_TABLE.sql
@@ -1,8 +1,8 @@
 CREATE TABLE episodes (
-     id INT NOT NULL AUTO_INCREMENT,
-     show_id INT NOT NULL,
-     season_number INT NOT NULL,
-     episode_number INT NOT NULL,
+     id BIGINT NOT NULL AUTO_INCREMENT,
+     show_id BIGINT NOT NULL,
+     season_number BIGINT NOT NULL,
+     episode_number BIGINT NOT NULL,
      PRIMARY KEY (id),
      INDEX episode_shows_ind (show_id),
     FOREIGN KEY (show_id)
